+++ plg_pickone/pickone.php
@@ -86,7 +86,7 @@ class plgSurveyPickone {
                 unset($old_sf_hid_field_ids[$i]);
         }
         
-        if(count($old_sf_hid_field_ids)){
+        if(count($old_sf_hid_field_ids) && !$data['issave2copy']){
             $query = "DELETE FROM `#__survey_force_fields` WHERE `quest_id` = '".$data['qid']."' AND ( id IN ( ".implode(', ', $old_sf_hid_field_ids)." ) ".($other_option_cb != 2? ' OR is_main = 0 ': '')." )";
             $database->setQuery($query);
             $database->execute();
@@ -96,7 +96,7 @@ class plgSurveyPickone {
             
             $f_row = $sf_hid_fields[$i];
             $new_field = JTable::getInstance('Fields', 'SurveyforceTable', array());
-            if ($sf_hid_field_ids[$i] > 0 ) {
+            if ($sf_hid_field_ids[$i] > 0 && !$data['issave2copy']) {
                 $new_field->id = $sf_hid_field_ids[$i];
             }
             $new_field->quest_id = $data['qid'];
@@ -125,7 +125,7 @@ class plgSurveyPickone {
             $other_id = JFactory::getApplication()->input->get('other_op_id', 0);
 
             $new_field = JTable::getInstance('Fields', 'SurveyforceTable', array());
-            if ($other_id > 0 ) {
+            if ($other_id > 0  && !$data['issave2copy']) {
                 $new_field->id = $other_id;
             }
             $new_field->quest_id = $data['qid'];
